# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Exchange Back-Office API
  version: 0.1.0
  description: Document validation workflow for exchange operations
servers:
  - url: http://localhost:3001/api
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /documents/upload:
    post:
      summary: Upload a document
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Document created, pipeline started
  /documents:
    get:
      summary: List documents
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: docType
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Paginated document list
  /documents/{id}:
    get:
      summary: Get document detail with reviews
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Document with extracted fields, validation errors, reviews
  /documents/{id}/review:
    post:
      summary: Submit review decision
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                decision: { type: string, enum: [APPROVED, REJECTED, CHANGES_REQUESTED] }
                comment: { type: string }
              required: [decision]
      responses:
        "200":
          description: Review recorded
...